Add files via upload
I developed a prediction model to address the challenges of a highly imbalanced credit card fraud detection dataset, which consists of 284,807 total samples, with only 492 instances classified as fraud. Despite the significant class imbalance, your model achieved impressive performance metrics, attaining a recall score of 0.82 and a precision score of 0.84 for the fraud class. These results indicate that the model is effective in identifying fraudulent transactions while maintaining a strong balance between sensitivity and specificity.